delete search.py, move astar to base_util
